rename uint -> size_t {file CType} add base class alloc

+++ include/CCompare.h
@@ -4,23 +4,28 @@
 
 namespace PetLib {
 
+    //////////////////////////////////////////
+    // LT < RT          return -1
+    // LT > RT          return  1
+    // LT == RT       return  0
+    //////////////////////////////////////////
     template< class type >
     int ARRMinCompare( const type& LT, const type& RT)
     {
-        uint Lsize = LT.size();
-        uint Rsize = RT.size();
-        uint smoll = ( Lsize < Rsize ) ? Lsize : Rsize;
+        size_t Lsize = LT.size();
+        size_t Rsize = RT.size();
+        size_t smoll = ( Lsize < Rsize ) ? Lsize : Rsize;
 
-        for(uint index = 0; index < smoll; ++index) 
+        for( size_t index = 0; index < smoll; ++index ) 
         {
-            if( LT[index] < RT[index] ) return -1; // left < right | -1
-            if( LT[index] > RT[index] ) return 1; // left > right | 1
+            if( LT[index] < RT[index] ) return -1;  // left < right | -1
+            if( LT[index] > RT[index] ) return 1;   // left > right |  1
         }
 
-        if( Lsize < Rsize ) return -1; // left < right | -1
-        if( Lsize > Rsize ) return 1; // left > right | 1
+        if( Lsize < Rsize ) return -1;  // left < right | -1
+        if( Lsize > Rsize ) return 1;   // left > right |  1
 
-        return 0; // left == right | 0
+        return 0;                       // left == right | 0
     }
 
 }; //PetLib
